backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic-settings==2.7.1
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.10.1
python-multipart==0.0.20
httpx==0.28.1
docker==7.1.0
tomlkit==0.13.2
